Initial CommonReturnType example.
This commit exhibits an example of where returning different types of values
from a method can cause confusing results to your consumers.

In the first example, `CollectionWithNonstandardReturnTypes` our `findByVal` is
a direct mapping to PHP's `array_search` which will:
> Return the key for needle if it is found in the array, FALSE otherwise.

Since our `test.php` script is using weak comparisons, if we try to detect if
a search term is the first value in the array, we will mistakenly fall into the
success condition when the array does not even contain this value.

This is then contrasted with, `CollectionWithStandardReturnTypes`, where we
throw a `ValueNotFoundException` for this case, which fails in a much more
obvious manner.

Example output:
```
$ php CommonReturnType/test.php
Search term 'm14t' is the first value in our collection!

=== vs ===

PHP Fatal error:  Uncaught exception 'm14t\ValueNotFoundException' in CommonReturnType/CollectionWithStandardReturnTypes.php:22
Stack trace:
  thrown in CommonReturnType/CollectionWithStandardReturnTypes.php on line 22

```