Nieve fix to the CommonReturnType example.
An easy fix to the previous example would simply be to use strict triple equals comparisons.

While this would prevent us from errouniously entering the success block of this example, it still lends itself to a bug where it is presumed that if the key is not the first element, that it is still present in the collection, when in fact, it may not be present at all.

Example output:
```
Search term 'm14t' NOT is the first value in our collection!

=== vs ===

PHP Fatal error:  Uncaught exception 'm14t\ValueNotFoundException' in CommonReturnType/CollectionWithStandardReturnTypes.php:22
Stack trace:
  thrown in CommonReturnType/CollectionWithStandardReturnTypes.php on line 22
```