Final fix for the CommonReturnType example.
The final solution is to add another elseif block to our nieve example, or to try and catch the error that was being thrown all along.

I propose that both of these solutions have the same complexity, but that the bugs were obvious much earlier when using exceptions, and when this 'edge case' did crop up would have failed exactly where the error was occuring and not at some later point in our business logic. This transparency makes our class easier to consume with a clearer communication of what is going on, without needing to read the documentation to find out what 'false' means.

Example output:
```
Search term 'm14t' not even in our collection!

=== vs ===

Search term 'm14t' not even in our collection!
```