Performance test for CommonReturnType example.
Some may argue that throwing exceptions is less ideal because they are slower (by about 65% according to this test), but if we are throwing an exception it is much more clear that this is the uncommon code path, it should not occur all that often, and when it does, it is both more clear that we are in the uncommon code path (or that this path even exists), and only ~0.00002 seconds longer. It is unlikely that this is the performance hit that is going to make or break your applications, and is dramatically outweighed by the odds of it preventing a bug from being created.